+++ src/main/java/com/moeasy/moeasy/dto/quesiton/PatchQuestionTitleResponseDto.java
@@ -1,6 +1,7 @@
 package com.moeasy.moeasy.dto.quesiton;
 
 
+import com.moeasy.moeasy.domain.question.Question;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -18,4 +19,11 @@ public class PatchQuestionTitleResponseDto {
 
   @Schema(description = "수정된 title 제목", example = "비사이드에 대한 설문 결과")
   private String title;
+
+  public static PatchQuestionTitleResponseDto from(Question question) {
+    return PatchQuestionTitleResponseDto.builder()
+        .id(question.getId())
+        .title(question.getTitle())
+        .build();
+  }
 }
